prune unused python code
