+++ rpsl_lexer/__init__.py
@@ -1,0 +1,2 @@
+def red(string: str) -> str:
+    return f"\033[91m{string}\033[0m"
